fix: do not restart api with every prediction

+++ src/dao/CreditCardDefault.py
@@ -1,5 +1,6 @@
+import os
 import pandas as pd
-from settings import PATH_CREDIT_CARD_DATASET
+from settings import PATH_CREDIT_CARD_DATASET, PATH_SAVE_PREDICTIONS
 
 def load_creditcard_dataset():
     try:
@@ -11,13 +12,26 @@ def load_creditcard_dataset():
         raise Exception(e)
 
 def write_predictions(df_predictions: pd.DataFrame):
+    file = f'creditcard_predictions.csv'
     try:
-        df_predictions.to_csv(
-            f'{PATH_CREDIT_CARD_DATASET}creditcard_predictions.csv',
-            index=False,
-            decimal='.',
-            sep=',',
-            mode='a'
-        )
+        if os.path.exists(PATH_SAVE_PREDICTIONS) == False:
+            os.mkdir(PATH_SAVE_PREDICTIONS)
+            df_predictions.to_csv(
+                f'{PATH_SAVE_PREDICTIONS}{file}',
+                index=False,
+                decimal='.',
+                sep=',',
+                mode='w'
+            )
+        else:
+            df_predictions.to_csv(
+                f'{PATH_SAVE_PREDICTIONS}{file}',
+                index=False,
+                decimal='.',
+                sep=',',
+                mode='a',
+                header=False
+            )
+
     except Exception as e:
         raise Exception(e)
